+++ playbook/actions/ai.go
@@ -6,13 +6,10 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"os"
-	"path/filepath"
 	"strings"
 	"text/template"
 
 	"github.com/flanksource/artifacts"
-	pkgConnection "github.com/flanksource/duty/connection"
 	"github.com/flanksource/duty/context"
 	"github.com/flanksource/duty/models"
 	"github.com/flanksource/duty/query"
@@ -29,8 +26,6 @@ import (
 	"github.com/flanksource/incident-commander/events"
 	"github.com/flanksource/incident-commander/llm"
 	llmContext "github.com/flanksource/incident-commander/llm/context"
-	"github.com/flanksource/incident-commander/pkg/clients/git"
-	"github.com/flanksource/incident-commander/pkg/clients/git/connectors"
 	"github.com/flanksource/incident-commander/utils"
 )
 
@@ -127,17 +122,6 @@ type childRunResultContext struct {
 func (t *aiAction) Run(ctx context.Context, spec v1.AIAction) (*AIActionResult, error) {
 	var result AIActionResult
 
-	// Resolve skill paths for the LLM skills middleware. Local paths are passed
-	// through directly; git paths are resolved inside the cloned repository.
-	var skillPaths []string
-	if len(spec.Skills) > 0 {
-		paths, err := resolveSkillPaths(ctx, spec.Skills)
-		if err != nil {
-			return nil, fmt.Errorf("failed to resolve skills: %w", err)
-		}
-		skillPaths = paths
-	}
-
 	knowledgebase, prompt, err := buildPrompt(ctx, spec.Prompt, spec.LLMContextRequest)
 	if err != nil {
 		return nil, fmt.Errorf("failed to form prompt: %w", err)
@@ -199,7 +183,15 @@ func (t *aiAction) Run(ctx context.Context, spec v1.AIAction) (*AIActionResult, 
 		}
 	}
 
-	llmConf := llm.Config{AIActionClient: spec.AIActionClient, SkillPaths: skillPaths}
+	skillPaths, err := resolveSkillPaths(ctx, spec.Skills)
+	if err != nil {
+		return nil, fmt.Errorf("failed to resolve skills: %w", err)
+	}
+
+	llmConf := llm.Config{
+		AIActionClient: spec.AIActionClient,
+		SkillPaths:     skillPaths,
+	}
 
 	// Use diagnosis schema when the requested output format needs it.
 	for _, format := range spec.Formats {
@@ -207,6 +199,7 @@ func (t *aiAction) Run(ctx context.Context, spec v1.AIAction) (*AIActionResult, 
 			llmConf.ResponseFormat = llm.ResponseFormatDiagnosis
 			break
 		}
+	}
 	}
 
 	// Resolve custom output schema if provided
@@ -497,126 +490,4 @@ func resolveOutputSchema(ctx context.Context, schema v1.AIOutputSchema) (string,
 	}
 
 	return "", nil
-}
-
-// cloneGitRepo resolves a git connection and clones the repo, returning the worktree root path.
-func cloneGitRepo(ctx context.Context, connectionRef, branch string) (string, error) {
-	conn, err := pkgConnection.Get(ctx, connectionRef)
-	if err != nil {
-		return "", fmt.Errorf("failed to get git connection %q: %w", connectionRef, err)
-	} else if conn == nil {
-		return "", fmt.Errorf("git connection %q not found", connectionRef)
-	}
-
-	spec := &connectors.GitopsAPISpec{
-		Repository: conn.URL,
-		Base:       "main",
-		Branch:     "main",
-	}
-
-	if branch != "" {
-		spec.Base = branch
-		spec.Branch = branch
-	}
-
-	switch conn.Type {
-	case models.ConnectionTypeGithub, models.ConnectionTypeGitlab, models.ConnectionTypeAzureDevops:
-		spec.AccessToken = conn.Password
-	case models.ConnectionTypeHTTP:
-		spec.User = conn.Username
-		spec.Password = conn.Password
-	case models.ConnectionTypeGit:
-		spec.User = conn.Username
-		spec.Password = conn.Password
-		spec.SSHPrivateKey = conn.Certificate
-		spec.SSHPrivateKeyPassword = conn.Password
-	default:
-		return "", fmt.Errorf("unsupported connection type %q", conn.Type)
-	}
-
-	_, workTree, err := git.Clone(ctx, spec)
-	if err != nil {
-		return "", fmt.Errorf("failed to clone repository: %w", err)
-	}
-
-	return workTree.Filesystem.Root(), nil
-}
-
-// safeResolvePath validates that filePath does not escape root via traversal.
-func safeResolvePath(root, filePath string) (string, error) {
-	if filepath.IsAbs(filePath) {
-		return "", fmt.Errorf("absolute paths are not allowed: %q", filePath)
-	}
-
-	joined := filepath.Join(root, filepath.Clean(filePath))
-	resolved, err := filepath.EvalSymlinks(joined)
-	if err != nil {
-		return "", fmt.Errorf("failed to resolve path %q: %w", filePath, err)
-	}
-
-	rel, err := filepath.Rel(root, resolved)
-	if err != nil || rel == ".." || strings.HasPrefix(rel, ".."+string(os.PathSeparator)) {
-		return "", fmt.Errorf("path %q resolves outside the repository root", filePath)
-	}
-
-	return resolved, nil
-}
-
-// safeReadFile validates that filePath does not escape root via traversal,
-// then reads and returns the file content.
-func safeReadFile(root, filePath string) ([]byte, error) {
-	resolved, err := safeResolvePath(root, filePath)
-	if err != nil {
-		return nil, err
-	}
-
-	return os.ReadFile(resolved)
-}
-
-// loadFileFromGit clones a git repo and reads a file at the given path.
-func loadFileFromGit(ctx context.Context, connectionRef, filePath, branch string) (string, error) {
-	root, err := cloneGitRepo(ctx, connectionRef, branch)
-	if err != nil {
-		return "", err
-	}
-
-	content, err := safeReadFile(root, filePath)
-	if err != nil {
-		return "", fmt.Errorf("failed to read file %q: %w", filePath, err)
-	}
-
-	return string(content), nil
-}
-
-// resolveSkillPaths resolves local and git skill paths for the LLM skills middleware.
-func resolveSkillPaths(ctx context.Context, skills []v1.AISkill) ([]string, error) {
-	paths := make([]string, 0, len(skills))
-
-	for i, skill := range skills {
-		if strings.TrimSpace(skill.Path) == "" {
-			return nil, fmt.Errorf("skill[%d]: path is required", i)
-		}
-
-		if skill.Connection != "" {
-			root, err := cloneGitRepo(ctx, skill.Connection, skill.Branch)
-			if err != nil {
-				return nil, fmt.Errorf("skill[%d]: %w", i, err)
-			}
-
-			resolved, err := safeResolvePath(root, skill.Path)
-			if err != nil {
-				return nil, fmt.Errorf("skill[%d] %q: %w", i, skill.Path, err)
-			}
-
-			paths = append(paths, resolved)
-			continue
-		}
-
-		if _, err := os.Stat(skill.Path); err != nil {
-			return nil, fmt.Errorf("skill[%d] %q: %w", i, skill.Path, err)
-		}
-		paths = append(paths, skill.Path)
-	}
-
-	return paths, nil
 }
